Moved CollideShapeVsShapePerLeaf function to Jolt library
